Clean up the logger
